requirements.txt
pyyaml
pydantic
watchdog
pillow
pytesseract
mss
